pyproject.toml
[project]
name = "hogutils"
version = "0.1.0"
description = "Utility scripts for manipulating Descent 3 HOG files"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
]
